+++ lib/primer/classify/cache.rb
@@ -96,12 +96,7 @@ module Primer
           )
 
           preload(
-            keys: [Primer::Classify::HEIGHT_KEY],
-            values: [:fit]
-          )
-
-          preload(
-            keys: [Primer::Classify::WIDTH_KEY],
+            keys: [Primer::Classify::HEIGHT_KEY, Primer::Classify::WIDTH_KEY],
             values: [:fit, :full]
           )