+++ test/lib/classify_test.rb
@@ -637,7 +637,9 @@ class PrimerClassifyTest < Minitest::Test
   def test_width_and_height
     assert_generated_class("width-fit",   { width: :fit })
     assert_generated_class("width-full",  { width: :full })
+
     assert_generated_class("height-fit",  { height: :fit })
+    assert_generated_class("height-full",  { height: :full })
   end
 
   def test_flex_grow