+++ CHANGELOG.md
@@ -40,7 +40,7 @@ The category for changes related to documentation, testing and tooling. Also, fo
 
     *Manuel Puyol*
 
-* Add support for width: :full to System Arguments.
+* Add support for `width: :full` and `height: :full` to System Arguments.
 
     *Joel Hawksley*